+++ app/src/main/java/com/travistorres/moviescout/utils/moviedb/MovieDbRequester.java
@@ -5,9 +5,16 @@
 package com.travistorres.moviescout.utils.moviedb;
 
 import android.content.Context;
-import android.os.AsyncTask;
+import android.content.res.Resources;
+import android.os.Bundle;
+import android.support.v4.app.FragmentActivity;
+import android.support.v4.app.LoaderManager;
+import android.support.v4.content.AsyncTaskLoader;
+import android.support.v4.content.Loader;
+import android.text.TextUtils;
 import android.widget.Toast;
 
+import com.travistorres.moviescout.R;
 import com.travistorres.moviescout.utils.moviedb.listeners.MovieClickedListener;
 import com.travistorres.moviescout.utils.moviedb.listeners.MovieDbNetworkingErrorHandler;
 import com.travistorres.moviescout.utils.moviedb.adapters.MovieListAdapter;
@@ -28,36 +35,43 @@ import java.net.URL;
  * Movie DB API.
  *
  * @author Travis Anthony Torres
- * @version February 21, 2017
+ * @version v1.2.0 (March 28, 2017)
  */
 
-public class MovieDbRequester {
+public class MovieDbRequester
+        implements LoaderManager.LoaderCallbacks<Movie[]> {
     /*
      *  Error message to display when no network could be reached.
      *
      */
     public final static String NO_NETWORK_ERROR_MESSAGE = "Unable to access Network Resource";
+    public final String MOVIE_REQUEST_URL_EXTRA;
 
     private MovieDbNetworkingErrorHandler errorHandler;
-    private Context context;
+    private FragmentActivity parentActivity;
     private MovieListAdapter movieAdapter;
     private int currentPage;
     private int totalPages;
     private int totalMovies;
     private MovieSortType sortType;
+    private String versionThreeApiKey;
+    private String versionFourApiKey;
 
     /**
      * Constructs a new Request object that will queried.
      *
-     * @param appContext
+     * @param parent
      * @param networkHandler
      * @param clickListener
      */
-    public MovieDbRequester(Context appContext, MovieDbNetworkingErrorHandler networkHandler, MovieClickedListener clickListener) {
-        context = appContext;
+    public MovieDbRequester(FragmentActivity parent, MovieDbNetworkingErrorHandler networkHandler, MovieClickedListener clickListener) {
+        parentActivity = parent;
         errorHandler = networkHandler;
         movieAdapter = new MovieListAdapter(clickListener, this);
         sortType = MovieSortType.MOST_POPULAR;
+
+        //  specifies the extra to use for acquiring the resource url
+        MOVIE_REQUEST_URL_EXTRA = parentActivity.getString(R.string.movie_request_url_extra);
 
         reset();
     }
@@ -69,6 +83,35 @@ public class MovieDbRequester {
      */
     public MovieListAdapter getAdapter() {
         return movieAdapter;
+    }
+
+    /**
+     * Specifies the api keys to use for accessing resources from the Movie DB.
+     *
+     * @param versionThreeKey
+     * @param versionFourKey
+     */
+    public void setApiKeys(String versionThreeKey, String versionFourKey) {
+        setVersionThreeApiKey(versionThreeKey);
+        setVersionFourApiKey(versionFourKey);
+    }
+
+    /**
+     * Specifies the key to use for accessing Version 3 API features.
+     *
+     * @param versionThreeKey
+     */
+    public void setVersionThreeApiKey(String versionThreeKey) {
+        versionThreeApiKey = versionThreeKey;
+    }
+
+    /**
+     * Specifies the key to use for accessing Version 4 API features.
+     *
+     * @param versionFourKey
+     */
+    public void setVersionFourApiKey(String versionFourKey) {
+        versionFourApiKey = versionFourKey;
     }
 
     /**
@@ -98,7 +141,17 @@ public class MovieDbRequester {
     public void requestNext() {
         if (hasNextPage()) {
             URL url = getCurrentRequestUrl();
-            new NetworkingTask().execute(url);
+            Bundle requestUrlBundle = new Bundle();
+            requestUrlBundle.putString(MOVIE_REQUEST_URL_EXTRA, url.toString());
+
+            //  retrieves the loader that will lookup the movies
+            Resources resources = parentActivity.getResources();
+            int loaderKey = resources.getInteger(R.integer.movie_db_requester_loader_manager_id);
+            LoaderManager loaderManager = parentActivity.getSupportLoaderManager();
+            Loader<Movie[]> movieLoader = loaderManager.getLoader(loaderKey);
+            loaderManager.restartLoader(loaderKey, requestUrlBundle, this);
+
+            //  update the page index
             ++currentPage;
         }
     }
@@ -109,8 +162,8 @@ public class MovieDbRequester {
      * @return The URL to acquire the next set of results.
      */
     public URL getCurrentRequestUrl() {
-        MovieDbUrlManager urlManager = new MovieDbUrlManager(context);
-        URL url = urlManager.getSortedMoveListUrl(sortType, currentPage);
+        MovieDbUrlManager urlManager = new MovieDbUrlManager(parentActivity);
+        URL url = urlManager.getSortedMoveListUrl(sortType, currentPage, versionThreeApiKey);
 
         return url;
     }
@@ -125,82 +178,131 @@ public class MovieDbRequester {
     }
 
     /**
-     * NetworkingTask
+     * Creates a thread that will issue a network requests for all of the Movies.
      *
-     * A utility class which will run a sub-process that queries a list of movies over the network.
-     * The process runs separately from the MainActivity to keep the app responsive during times of
-     * low/non-existent network connectivity.
+     * @param id
+     * @param args
      *
+     * @return LoaderManager responsible for requesting movie resources from the server
      */
+    @Override
+    public Loader<Movie[]> onCreateLoader(int id, final Bundle args) {
+        return new AsyncTaskLoader<Movie[]>(parentActivity) {
+            /**
+             * Displays the loading wheel so that users are made aware that a requests is occurring
+             * in the background.
+             *
+             */
+            @Override
+            protected void onStartLoading() {
+                super.onStartLoading();
 
-    private class NetworkingTask extends AsyncTask<URL, Void, Movie[]> {
-        /**
-         * Allow users to see the loading indicator wheel.
-         *
-         */
-        @Override
-        protected void onPreExecute() {
-            super.onPreExecute();
-
-            errorHandler.beforeNetworkRequest();
-        }
-
-        /**
-         * Request a list of movies from the network resource.
-         *
-         * @param urls List of urls for acquiring the movies.  We only need one.
-         *
-         * @return List of movies that were queried.
-         */
-        @Override
-        protected Movie[] doInBackground(URL... urls) {
-            if (urls.length <= 0) {
-                return null;
-            }
-
-            URL url = urls[0];
-            Movie[] movieList = null;
-
-            try {
-                String json = NetworkManager.request(url);
-                if (json != null) {
-                    movieList = MovieDbParser.retrieveMovieList(json, context);
-                    totalMovies = MovieDbParser.acquireTotalResults(json);
-                    totalPages = MovieDbParser.acquireTotalPages(json);
+                if (args == null) {
+                    return;
                 }
-            } catch (IOException e) {
-                e.printStackTrace();
-            } catch (HttpConnectionTimeoutException e) {
-                e.printStackTrace();
-            } catch (HttpPageNotFoundException e) {
-                e.printStackTrace();
-                errorHandler.onPageNotFound();
-            } catch (HttpUnauthorizedException e) {
-                e.printStackTrace();
-                errorHandler.onUnauthorizedAccess();
-            } catch (NetworkingException e) {
-                e.printStackTrace();
-                errorHandler.onGeneralNetworkingError();
+
+                //  displays the loading wheel
+                errorHandler.beforeNetworkRequest();
+
+                //  forces content to display
+                forceLoad();
             }
 
-            return movieList;
-        }
+            /**
+             * Issues a network request for all movie information from MovieDbApi.com.
+             *
+             * @return List of movies obtained from the server.
+             */
+            @Override
+            public Movie[] loadInBackground() {
+                String urlString = null;
+                Movie[] movieList = null;
 
-        /**
-         * Stores the list of movies in the adapter so they can be efficiently rendered onto the
-         * app.
-         *
-         * @param list acquired movies.
-         */
-        @Override
-        protected void onPostExecute(Movie[] list) {
-            errorHandler.afterNetworkRequest();
+                //  acquire the string value of the URL if one has been set
+                if (args.containsKey(MOVIE_REQUEST_URL_EXTRA)) {
+                    urlString = args.getString(MOVIE_REQUEST_URL_EXTRA);
+                    if (urlString == null || TextUtils.isEmpty(urlString)) {
+                        return null;
+                    }
+                } else {
+                    return null;
+                }
 
-            if (list != null) {
-                movieAdapter.setMoviesList(list);
-            } else {
-                Toast.makeText(context, NO_NETWORK_ERROR_MESSAGE, Toast.LENGTH_SHORT).show();
+                //  attempt to acquire the resource
+                try {
+                    URL url = new URL(urlString);
+                    String json = NetworkManager.request(url);
+                    if (json != null) {
+                        movieList = MovieDbParser.retrieveMovieList(json);
+                        totalMovies = MovieDbParser.acquireTotalResults(json);
+                        totalPages = MovieDbParser.acquireTotalPages(json);
+                    }
+                } catch (IOException e) {
+                    e.printStackTrace();
+                } catch (HttpConnectionTimeoutException e) {
+                    e.printStackTrace();
+                } catch (HttpPageNotFoundException e) {
+                    e.printStackTrace();
+                    errorHandler.onPageNotFound();
+                } catch (HttpUnauthorizedException e) {
+                    e.printStackTrace();
+                    errorHandler.onUnauthorizedAccess();
+                } catch (NetworkingException e) {
+                    e.printStackTrace();
+                    errorHandler.onGeneralNetworkingError();
+                }
+
+                return movieList;
             }
+        };
+    }
+
+    /**
+     * Specifies what to do with the movie resources that were obtained from the server.
+     *
+     * @param loader
+     * @param list
+     */
+    @Override
+    public void onLoadFinished(Loader<Movie[]> loader, Movie[] list) {
+        errorHandler.afterNetworkRequest();
+
+        //  add all of the movies to the list
+        if (list != null) {
+            movieAdapter.setMoviesList(list);
+        } else {
+            Toast.makeText(parentActivity, NO_NETWORK_ERROR_MESSAGE, Toast.LENGTH_SHORT).show();
         }
+    }
+
+    /**
+     * Specifies what to do when the loader is reset.
+     *
+     * @param loader
+     */
+    @Override
+    public void onLoaderReset(Loader<Movie[]> loader) {
+        //  Intentionally left blank
+    }
+
+    public int getCurrentPage() {
+        return currentPage;
+    }
+
+    public MovieSortType getSortType() {
+        return sortType;
+    }
+
+    public void setCurrentPage(int currentPage) {
+        this.currentPage = currentPage;
+    }
+
+    /**
+     * Retrieves the context which the Requester is acting on.
+     *
+     * @return calling context
+     */
+    public Context getContext() {
+        return parentActivity;
     }
 }